fix
Fixed grep error when parsing /var/log/auth.log; grep is treating auth.log as a binary file for some reason. This might be a system specific issue as it never happened before; using 'grep -a' to parse auth.log is a workaround that forces grep to treat the file as text.